# server/src/docs/swagger.yaml
openapi: 3.0.2
info:
    title: Library-API
    version: 1.0.0
    description: '### This is a Swagger UI for our Library API'
paths:
...
    '/books/{id}':
        get:
            tags:
                - books
            parameters:
                -
                    examples:
                        id:
                            value: '1'
                    name: id
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/book'
                            examples:
                                Book:
                                    value:
                                        id: 1
                                        title: Title1
                                        description: >-
                                            Romeo and Juliet, play by William Shakespeare, written about
                                            1594–96 and first published in an unauthorized quarto in 1597. An
                                            authorized quarto appeared in 1599, substantially longer and more
                                            reliable.
                                        author: William Shakespeare2
                                        releasedate: '1597-01-01T00:00:00.000Z'
                                        language: English
                                        booklength: 480
                                        price: 5.99
                                        createdAt: '2023-10-24T18:55:57.000Z'
                                        updatedAt: '2023-10-26T11:46:19.000Z'
                    description: >-
                        Returns book by id (id, title, author, description, releasedate, booklength, language,
                        price, createdAt, updatedAt)
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Not found.:
                                    value:
                                        error: book not found.
                    description: Returns error
            summary: Get one book
            description: Returns book by id
        put:
            requestBody:
                description: example request body data
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/book'
                        examples:
                            example request body data:
                                value:
                                    title: Middlemarch 2
                                    author: Eliot
                                    description: >-
                                        Middlemarch is a novel that explores the dynamics of marriage and
                                        social class in 19th-century England.
                                    price: 3.95
                            example invalid request body data:
                                value: {}
                required: true
            tags:
                - books
            parameters:
                -
                    examples:
                        id:
                            value: '4'
                    name: id
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '200':
                    headers:
                        Location:
                            schema:
                                type: string
                            examples:
                                Location of book:
                                    value: 'http://localhost:3000/books/4'
                    description: OK
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                field is required:
                                    value:
                                        error: At least one field is required.
                                Release Date is invalid:
                                    value:
                                        error: releasedate invalid.
                    description: some fields are missing
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Not found:
                                    value:
                                        error: book not found.
                    description: book not found
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Update failed:
                                    value:
                                        error: could not update book
                    description: Internal server error
            summary: Update book
            description: Updates book by id
        delete:
            tags:
                - books
            parameters:
                -
                    examples:
                        id:
                            value: '4'
                    name: id
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '204':
                    description: No content
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Book not found.:
                                    value:
                                        error: book not found.
                    description: Book not found
                '409':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                'Could not delete book, Order exists':
                                    value:
                                        error: book could not be deleted due to an existing order
                    description: book could not be deleted due to an existing order
            summary: Delete book by id
            description: Deletes book using its id
...
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
...
        book:
            title: Root Type for book
            description: book entity
            type: object
            properties:
                id:
                    format: int32
                    type: integer
                title:
                    type: string
                author:
                    type: string
                description:
                    type: string
                releasedate:
                    type: string
                booklength:
                    format: int32
                    type: integer
                language:
                    type: string
                price:
                    format: double
                    type: number
                pdf:
                    type: string
            example:
                id: 1
                title: Romeo and Juliet
                author: William Shakespeare
                description: >-
                    Romeo and Juliet, play by William Shakespeare, written about 1594–96 and first published
                    in an unauthorized quarto in 1597. An authorized quarto appeared in 1599, substantially
                    longer and more reliable.
                releasedate: '1597'
                booklength: 480
                language: English
                price: 5.99
                pdf: '?'
...